fix the clueboard->clueboard_66 rename
