--- keyboards/clueboard_66/clueboard.c
@@ -1,1 +1,0 @@
-#include "clueboard.h"
