+++ keyboards/clueboard_66/clueboard_66.c
@@ -1,0 +1,1 @@
+#include "clueboard_66.h"
