+++ keyboards/clueboard_66/rev1/rev1.h
@@ -1,7 +1,7 @@
 #ifndef REV1_H
 #define REV1_H
 
-#include "../clueboard.h"
+#include "../clueboard_66.h"
 
 /* Clueboard matrix layout
  * ,-----------------------------------------------------------.  ,---.
